Rename: 파일들이나 fun 이름을 convention에 맞춰서 page -> screen 으로 수정, 조교님 피드백을 반영하여  search -> social로 수정
